Fix hash checks when we don't push root references
RepoRefsCheckHash currently checks all references but there are cases
where we only push /refs. And this is actually the main logic when the
tool pushes to the mirror repository. So in that case, checking for
other refs, for example HEAD, makes no sense. This patch adds a prefix
parameter that we can use to only check references prefixed with a
specific string. With this change, after we push (mirror push) to a
repository, we can verify only the references refs/* - which are the
ones we actually push.

This change only affects/improves testing.

Signed-off-by: Andrei Gherzan <andrei@gherzan.com>